requirements.txt
chainer==4.1.0
chainermn==1.3.0
cupy==4.1.0
pillow==3.3.1
scipy==1.1.0
mpi4py==3.0.0
pyyaml==3.12
